fix behind msca 500ing, typeOfApplication, and member-selection
